+++ sv-parser-parser/src/general/compiler_directives.rs
@@ -478,6 +478,9 @@ pub(crate) fn else_group_of_lines(s: Span) -> IResult<Span, ElseGroupOfLines> {
 pub(crate) fn source_description(s: Span) -> IResult<Span, SourceDescription> {
     alt((
         map(comment, |x| SourceDescription::Comment(Box::new(x))),
+        map(string_literal, |x| {
+            SourceDescription::StringLiteral(Box::new(x))
+        }),
         source_description_not_directive,
         map(compiler_directive, |x| {
             SourceDescription::CompilerDirective(Box::new(x))
@@ -489,7 +492,7 @@ pub(crate) fn source_description(s: Span) -> IResult<Span, SourceDescription> {
 #[packrat_parser]
 pub(crate) fn source_description_not_directive(s: Span) -> IResult<Span, SourceDescription> {
     let (s, a) = many1(alt((
-        is_not("`/"),
+        is_not("`/\""),
         terminated(tag("/"), peek(not(alt((tag("/"), tag("*")))))),
     )))(s)?;