+++ sv-parser-pp/testcases/test5.sv
@@ -1,0 +1,10 @@
+module a;
+`define HI Hello
+`define LO "`HI, world"
+`define H(x) "Hello, x"
+initial begin
+$display("`HI, world");
+$display(`LO);
+$display(`H(world));
+end
+endmodule